refactor(commands): change command prefix from # to @ (#9)
We want to be able to use gsh as a drop-in replacement for sh,
bash, zsh, etc. Unfortunately, the "#" is strongly associated
with comments in Posix shells, and that interferes with being
able to write scripts which mix agent actions with shell commands
in gsh. It's also a problem when trying to run scripts originally
intended for sh or bash.

This change basically just swaps it so that "@" is the magic word
for agent stuff instead of "#".

- Update all command prefixes from # to @ throughout codebase
- Change #/ macros to @/ macros in documentation and code
- Update #! subagent commands to @! subagent commands
- Modify #@ subagent completion to @ subagent completion
- Update core shell logic to recognize @ prefix instead of #
- Maintain all existing functionality with new syntax
- Update all test cases and integration tests to use new syntax